+++ RoaringBitmap/src/main/java/org/roaringbitmap/bsi/buffer/BitSliceIndexBase.java
@@ -446,7 +446,7 @@ public class BitSliceIndexBase {
         long count = foundSet.getLongCardinality();
 
         Long sum = IntStream.range(0, this.bitCount())
-                .mapToLong(x -> (1 << x) * ImmutableRoaringBitmap.andCardinality(this.bA[x], foundSet))
+                .mapToLong(x -> (long) (1 << x) * ImmutableRoaringBitmap.andCardinality(this.bA[x], foundSet))
                 .sum();
 
         return Pair.newPair(sum, count);
